requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
pdfplumber>=0.10.3
pypdf>=4.0.0
python-docx>=1.1.0
plotly>=5.20.0
pandas>=2.0.0
requests>=2.31.0
streamlit-lottie>=0.0.5
Pillow>=10.0.0
reportlab>=4.0.0
fpdf2>=2.7.0
openpyxl>=3.1.2
numpy>=1.26.0
